feat: Task 6
