--- src/main/java/com/rakovets/course/java/core/practice/enum_types/FavoriteSeason.java
@@ -1,28 +1,0 @@
-package com.rakovets.course.java.core.practice.enum_types;
-
-public interface SeasonDescription {
-    default String printFavoriteSeason(Seasons favoriteSeason) {
-        String result;
-
-        switch (favoriteSeason) {
-            case SPRING:
-                result = "I like Spring!";
-                break;
-            case SUMMER:
-                result = "I like Summer!";
-                break;
-            case AUTUMN:
-                result = "I like Autumn!";
-                break;
-            case WINTER:
-                result = "I like Winter!";
-                break;
-            default:
-                result = "There is no such season!";
-        }
-
-        return result;
-    }
-
-    default String
-}
