+++ src/test/java/com/rakovets/course/java/core/practice/enum_types/MyFavoriteSeasonInfoTest.java
@@ -6,19 +6,19 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 public class MyFavoriteSeasonInfoTest {
-    static private MyFavoriteSeasonInfo myFavoriteSeasonInfo;
+    static private Season season;
     private String result;
     private String expected;
 
     @BeforeAll
     static void initAll() {
-        myFavoriteSeasonInfo = new MyFavoriteSeasonInfo(Seasons.SPRING);
+        season = new Season(Seasons.SPRING);
     }
 
     @DisplayName("Test favoriteSeasonInfo(), returns String with info about my favorite season, Spring")
     @Test
     void printFavoriteSeasonInfoTest() {
-        result = myFavoriteSeasonInfo.favoriteSeasonInfo();
+        result = season.favoriteSeasonInfo();
         expected = "Spring is a transition season between the winter and summer seasons. The days get \n" +
                 "longer and nights get shorter, the temperatures are milder, and flowers bloom. There is a warm \n" +
                 "wind in the spring season in the air. Spring typically begins in March and usually lasts until \n" +
@@ -30,7 +30,7 @@ public class MyFavoriteSeasonInfoTest {
     @DisplayName("Test printFavoriteSeason(), returns String 'I like Spring!'")
     @Test
     void printFavoriteSeasonTest() {
-        result = myFavoriteSeasonInfo.printFavoriteSeason(myFavoriteSeasonInfo.getFAVORITE_SEASON());
+        result = season.printFavoriteSeason(season.getFAVORITE_SEASON());
         expected = "I like Spring!";
 
         Assertions.assertEquals(expected, result);
